requirements.txt
psycopg2==2.8.5
discord.py==1.3.3
google==2.0.3